api/requirements.txt
# Django and DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Authentication
djangorestframework-simplejwt==5.3.0
boto3==1.34.0
botocore==1.34.0

# Database
psycopg2-binary==2.9.9

# Celery
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# WebSockets
channels==4.0.0
channels-redis==4.1.0

# PDF Generation
weasyprint==60.2
reportlab==4.0.7

# Payments
stripe==7.8.0

# File Storage
boto3==1.34.0

# Notifications
twilio==8.10.0
sendgrid==6.10.0

# Utilities
python-decouple==3.8
Pillow==10.1.0
python-dateutil==2.8.2

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Security
django-ratelimit==4.1.0
cryptography==41.0.8

# Monitoring
sentry-sdk==1.38.0
